Refactored search activity into a fragment to avoid having to manually add existing navigation elements (actionbar) and turned search layout into a linear layout to keep things aligned automatically
