+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LibraryActivity" />
 
         <activity android:name=".MainActivity" />
