--- app/src/main/res/layout/activity_library.xml
@@ -1,38 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              android:orientation="vertical"
-              android:padding="16dp">
-
-    <EditText
-            android:id="@+id/searchInput"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Search comics..." />
-
-    <Button
-            android:id="@+id/searchButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Search" />
-
-    <Button
-            android:id="@+id/loadMoreButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Load More"
-            android:layout_marginTop="16dp"/>
-
-    <Spinner
-            android:id="@+id/publisherFilter"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"/>
-
-    <Spinner
-            android:id="@+id/yearFilter"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"/>
-
-</LinearLayout>
