+++ app/src/main/res/layout/fragment_search.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:padding="16dp">
+
+    <LinearLayout
+            android:orientation="vertical"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+
+        <EditText
+                android:id="@+id/searchInput"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:hint="Enter comic title" android:minHeight="48dp"/>
+
+        <Spinner
+                android:id="@+id/publisherFilter"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content" android:minHeight="48dp"/>
+
+        <Spinner
+                android:id="@+id/yearFilter"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content" android:minHeight="48dp"/>
+
+        <Button
+                android:id="@+id/searchButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Search" />
+
+        <Button
+                android:id="@+id/loadMoreButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Load More" />
+
+    </LinearLayout>
+</ScrollView>
