+++ app/src/main/res/menu/main.xml
@@ -1,6 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item android:id="@+id/action_search"
+          android:title="Search"
+          app:showAsAction="always" android:icon="@android:drawable/ic_menu_search"/>
+
     <item android:id="@+id/exampleAddComic"
           android:title="E.g. Add Comic"
           app:showAsAction="never"/>
